ai_pipeline/requirements.txt
pandas
openai
schedule
python-dotenv
streamlit
